backend/requirements.txt
# Migrated from: TruthLens/requirements.txt
# Enhanced for FastAPI backend deployment

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Async HTTP Client
aiohttp==3.9.0
httpx==0.25.2

# AI & NLP
textblob==0.17.1
nltk==3.8.1
spacy==3.7.2

# Google Cloud Services
google-generativeai==0.3.2
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-cloud-language==2.11.1
google-auth==2.23.4

# Firebase
firebase-admin==6.2.0

# Security
cryptography==41.0.7
bcrypt==4.1.1
python-jose==3.3.0
passlib==1.7.4

# Environment & Config
python-dotenv==1.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Image Processing
Pillow==10.1.0
# opencv-python==4.8.1.78

# Utilities
python-dateutil==2.8.2
pytz==2023.3
faker==20.1.0
bleach==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring
structlog==23.2.0

# Production
gunicorn==21.2.0

# Uploads: multipart/form-data parsing for FastAPI UploadFile/Form (/api/upload)
python-multipart>=0.0.9